Add the C++11 features supported by Visual Studio 2015
Tested with the November 2014 Preview.

There are three C++11 features that aren't supported as of MSVC2015:
 - attributes
 - constexpr
 - uniform initialization

Microsoft recognizes that constexpr support in this compiler is partial,
so we don't enable it yet. Attributes are not implemented at all.

As for uniform initialization, Microsoft's feature list says
initalizer_list is supported, but we still see compile errors trying to
initialize arrays in constructor initialization lists.

Change-Id: Id686ccb6b4b2ecd63f16d308552174db1f05f2c5
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>