QDebug: Remove handling of FORCE_UREF define
This is the only place in Qt source code we use FORCE_UREF, and can
be traced back to the very first qt.git commit. In any case, it's
broken: It returns a reference to a local variable, since the debug
argument is _not_ a reference.

Using a reference both for the argument and the return value would
be actually the canonical solution, but that breaks with

QDebug << operator(QDebug, const QVariant &),

exported in QtCore. The C++ lookup rules apparently prefer this
overload then to be used for outputting containers ...

Change-Id: Iaf5e5dd89d4f3ebe6454eba219046b4f25b0d717
Reviewed-by: hjk <hjk121@nokiamail.com>