Update the Win32 Intel Compiler mkspec
Unlike MSVC, ICC is capable of selecting each of the processor feature
levels, so let's define the right macros.

Version 9.1 is really old and not supported, so we don't need to keep
the old workaround.

The compiler has been complaining that option -GX is deprecated and will
be removed, so update it to use the same as MSVC does.

Change-Id: I4158fcf2331c1d27462bb1cb19725c7136efab4a
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>