Remove the rcc output file if we failed to write it
This happens if the second pass failed to find the signature data in the
compiled output, which can happen if the compiler is not *actually*
compiling anything (LTO / LTCG mode).

If we left the unmodified file in the output, make would continue
running next time it was invoked, resulting in bad data.

Change-Id: I97f23a89a1ca1f8b8c449b0744de3f0c78daa706
Reviewed-by: hjk <hjk121@nokiamail.com>