Doc: update the list of typical values for xxxCpuArchitecture
If we wanted to make the list exhaustive for buildCpuArchitecture, we'd
add alpha, avr32, bfin, s390(x), and sh. I don't want to make it
exhaustive because that's cluttering the documentation and we'd have to
keep in sync with archdetect.cpp. And we can't make
currentCpuArchitecture exhaustive, since it depends on the output from
uname(2).

You can argue that neither IA-64, POWER, nor SPARC architectures are
typical these days...

Change-Id: I0a5310770947263e1bafd9443ea59420813c51a8
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Jake Petroules <jake.petroules@petroules.com>
Reviewed-by: Richard J. Moore <rich@kde.org>