Rework of QSysInfo OS and kernel version functions
This patch splits the OS product from the kernel name and versions.
Previously, it was a little confusing.

Thus, osType is split in two: kernelType() always returns the kernel
type, which includes "linux" for Android, "darwin" for iOS and OS X,
"qnx" for BlackBerry, and "windows" for all the Windows OSes. The Linux
distribution name and the visible product names are found in
productType(), with the respective versions in productVersion().

For an update system, applications probably want to send all four
tokens:
 - for Android, BlackBerry, iOS, OS X and Windows, the relevant
   information is productType and productVersion
 - for Linux distributions, in addition to productType and
   productVersion, the kernelType ("linux") is useful as a fallback and
   it's important for Debian/kFreeBSD and Debian/GNU
 - for other Unix systems, the relevant information is kernelType and
   kernelVersion

Change-Id: I66c36d5a4ddfad47babfc6744d752273705e7076
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>